refactor (validator): 입력값 검증 메소드 이름 inputValidate로 일괄 변경
