+++ src/test/java/lotto/validator/LottoValidatorTest.java
@@ -1,0 +1,5 @@
+package lotto.validator;
+
+class LottoValidatorTest {
+
+}
